refactor: extract refcount table types into meta/refcount.rs
Move RefTableEntry, RefTable, RefBlockEntry, RefBlock, their trait
impls and the three refcount unit tests into a new meta/refcount.rs
submodule. test_refcount_block2 moves with the code because it builds
RefBlockEntry via its private tuple field. meta/mod.rs is now reduced
to module declarations and re-exports. No logic changes.
